966.1 Components structure refactor
