+++ project/src/components/comment-form/comments-list.tsx
@@ -2,7 +2,7 @@ import {CommentItem} from './index';
 
 import {Comments} from '../../types/offers';
 
-import {MAX_COMMENT_COUNT} from '../../const';
+import {Comment} from '../../const';
 
 type CommentListProps = {
   comments: Comments;
@@ -13,7 +13,7 @@ function CommentsList({comments}: CommentListProps): JSX.Element {
   return (
     <ul className="reviews__list">
       {comments.map((comment, idx) => (
-        idx < MAX_COMMENT_COUNT
+        idx < Comment.MaxCount
           ? <CommentItem key={comment.id} comment={comment}/>
           : null
       ))}
