+++ project/src/components/layout/index.ts
@@ -1,2 +1,3 @@
+export {default as Layout} from './layout';
 export {default as Footer} from './footer/footer';
 export {default as Header} from './header/header';
