+++ project/src/components/main/index.ts
@@ -1,5 +1,4 @@
-export {default as CitiesList} from './cities-list';
-export {default as OfferCardsList} from './offer-cards-list';
-export {default as SortingForm} from './sorting-form';
-export {default as Premium} from './premium';
-export {default as OfferCard} from './offer-card';
+export {default as CitiesList} from './cities-list/cities-list';
+export {default as OfferCardsList} from './offer-cards-list/offer-cards-list';
+export {default as SortingForm} from './sorting-form/sorting-form';
+export {default as Premium} from './premium/premium';
