+++ project/src/components/nav/index.ts
@@ -1,3 +1,3 @@
 export {default as Nav} from './nav';
-export {default as SignOut} from './sign-out';
-export {default as SignIn} from './sign-in';
+export {default as SignOut} from './sign-out/sign-out';
+export {default as SignIn} from './sign-in/sign-in';
