+++ project/src/components/offer/index.ts
@@ -1,0 +1,3 @@
+export {default as OfferInsideItem} from './offer-inside-item/offer-inside-item';
+export {default as OfferNearList} from './offer-near-list/offer-near-list';
+export {Gallery, Property, Host, Reviews} from './internal';
