+++ project/src/components/offer/internal/gallery.tsx
