+++ project/src/components/offer/internal/index.ts
