+++ project/src/components/offer/internal/property.tsx
