+++ project/src/components/offer/offer-inside-item/offer-inside-item.tsx
