+++ project/src/components/offer/offer-near-list/offer-near-list.tsx
@@ -1,6 +1,6 @@
-import {OfferCard} from '../main';
+import OfferCard from '../../offer-card/offer-card';
 
-import {Offers} from '../../types/offers';
+import {Offers} from '../../../types/offers';
 
 type OfferNearCardProps = {
   offersNear: Offers;
