--- project/src/components/room/index.ts
@@ -1,5 +1,0 @@
-export {default as OfferInsideItem} from './features-inside';
-export {default as OfferNearsList} from './offer-near-list';
-export {CommentForm, CommentsList, CommentItem} from '../comment';
-export {Gallery, Property, Host, Reviews} from './internal';
-
