+++ project/src/pages/favorites-screen/favorites-screen.tsx
@@ -1,5 +1,5 @@
-import Layout from '../../components/layout/layout';
-import {FavoritesList} from '../../components/favorites';
+import {Layout} from '../../components/layout';
+import {Favorites} from '../../components/favorites';
 
 type FavoritesScreenProps = {
   onChangeCity: (city: string) => void;
@@ -13,7 +13,7 @@ function FavoritesScreen({onChangeCity}: FavoritesScreenProps): JSX.Element {
           <div className="page__favorites-container container">
             <section className="favorites">
               <h1 className="favorites__title">Saved listing</h1>
-              <FavoritesList onChangeCity={onChangeCity} />
+              <Favorites onChangeCity={onChangeCity} />
             </section>
           </div>
         </main>
