+++ project/src/pages/main-empty-screen/main-empty-screen.tsx
@@ -1,4 +1,4 @@
-import Layout from '../../components/layout/layout';
+import {Layout} from '../../components/layout';
 import {CitiesList} from '../../components/main';
 
 type MainEmptyScreenProps = {
